Convert `JavaLibraryPluginPerformanceTest`
