+++ .teamcity/performance-test-runtimes.csv
@@ -115,3 +115,4 @@ org.gradle.performance.regression.android.FasterIncrementalAndroidBuildsPerforma
 org.gradle.performance.regression.android.FasterIncrementalAndroidBuildsPerformanceTest;faster non-abi change (build comparison);santaTrackerAndroidJavaBuild;linux;398000
 org.gradle.performance.regression.android.FasterIncrementalAndroidBuildsPerformanceTest;faster abi change (build comparison);santaTrackerAndroidBuild;linux;795000
 org.gradle.performance.regression.android.FasterIncrementalAndroidBuildsPerformanceTest;faster abi change (build comparison);santaTrackerAndroidJavaBuild;linux;425000
+org.gradle.performance.experiment.java.JavaLibraryPluginPerformanceTest;java-library vs java;largeJavaMultiProject;linux;352000
