+++ subprojects/internal-performance-testing/src/main/groovy/org/gradle/performance/generator/TestProjects.groovy
@@ -1,0 +1,55 @@
+/*
+ * Copyright 2020 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.performance.generator
+
+class TestProjects {
+    static List<String> getProjectMemoryOptions(String testProject) {
+        def daemonMemory = determineDaemonMemory(testProject)
+        return ["-Xms${daemonMemory}", "-Xmx${daemonMemory}"]
+    }
+
+    private static String determineDaemonMemory(String testProject) {
+        switch (testProject) {
+            case 'bigCppApp':
+                return '256m'
+            case 'bigCppMulti':
+                return '1g'
+            case 'nativeDependents':
+                return '3g'
+            case 'bigNative':
+                return '1g'
+            case 'withVerboseJUnit':
+                return '256m'
+            case 'withVerboseTestNG':
+                return '256m'
+            case 'mediumSwiftMulti':
+                return '1G'
+            case 'bigSwiftApp':
+                return '1G'
+            case 'nativeMonolithic':
+                return '2500m'
+            case 'nativeMonolithicOverlapping':
+                return '2500m'
+            case 'mediumNativeMonolithic':
+                return '512m'
+            case 'smallNativeMonolithic':
+                return '512m'
+            default:
+                return JavaTestProject.projectFor(testProject).daemonMemory
+        }
+    }
+}
