+++ subprojects/performance/src/performanceTest/groovy/org/gradle/performance/experiment/java/JavaLibraryPluginPerformanceTest.groovy
@@ -22,15 +22,11 @@ import org.gradle.performance.categories.SlowPerformanceRegressionTest
 import org.gradle.performance.results.BaselineVersion
 import org.gradle.performance.results.CrossBuildPerformanceResults
 import org.junit.experimental.categories.Category
-import spock.lang.Unroll
-
-import static org.gradle.performance.generator.JavaTestProject.LARGE_JAVA_MULTI_PROJECT
 
 @Category(SlowPerformanceRegressionTest)
 class JavaLibraryPluginPerformanceTest extends AbstractCrossBuildPerformanceTest {
 
-    @Unroll
-    def "java-library vs java on #testProject"() {
+    def "java-library vs java"() {
         def javaLibraryRuns = "java-library-plugin"
         def javaRuns = "java-plugin"
         def compileClasspathPackaging = OperatingSystem.current().windows
@@ -38,17 +34,23 @@ class JavaLibraryPluginPerformanceTest extends AbstractCrossBuildPerformanceTest
         given:
         runner.testGroup = "java plugins"
         runner.buildSpec {
-            warmUpCount = warmUpRuns
-            invocationCount = runs
-            projectName(testProject.projectName).displayName(javaLibraryRuns).invocation {
-                tasksToRun("clean", "classes").args("-PcompileConfiguration", "-Dorg.gradle.java.compile-classpath-packaging=$compileClasspathPackaging").gradleOpts("-Xms${testProject.daemonMemory}", "-Xmx${testProject.daemonMemory}")
+            warmUpCount = 2
+            invocationCount = 3
+            displayName(javaLibraryRuns)
+            invocation {
+                tasksToRun("clean", "classes")
+                args("-PcompileConfiguration", "-Dorg.gradle.java.compile-classpath-packaging=$compileClasspathPackaging")
+                gradleOpts(runner.projectMemoryOptions)
             }
         }
         runner.baseline {
-            warmUpCount = warmUpRuns
-            invocationCount = runs
-            projectName(testProject.projectName).displayName(javaRuns).invocation {
-                tasksToRun("clean", "classes").args("-PcompileConfiguration", "-PnoJavaLibraryPlugin").gradleOpts("-Xms${testProject.daemonMemory}", "-Xmx${testProject.daemonMemory}")
+            warmUpCount = 2
+            invocationCount = 3
+            displayName(javaRuns)
+            invocation {
+                tasksToRun("clean", "classes")
+                args("-PcompileConfiguration", "-PnoJavaLibraryPlugin")
+                gradleOpts(runner.projectMemoryOptions)
             }
         }
 
@@ -63,11 +65,6 @@ class JavaLibraryPluginPerformanceTest extends AbstractCrossBuildPerformanceTest
         if (javaResults.significantlyFasterThan(javaLibraryResults, 0.95)) {
             throw new AssertionError(speedStats)
         }
-
-        where:
-        testProject                   | warmUpRuns | runs
-        LARGE_JAVA_MULTI_PROJECT      | 2          | 3
-        // HUGE_JAVA_MULTI_PROJECT    | 2          | 3
     }
 
     private static BaselineVersion buildBaselineResults(CrossBuildPerformanceResults results, String name) {
